add error for extra repos in keep
Since keep is an optional parameter, then we need to be mindful of the
fact that users who use this option intend for these specific
repositories to be archived. If there is a repository that does not
exist, it is better to stop early and alert the user instead of issuing
a warning and dropping the non-existent repository, which would eat up
bandwidth when there might be a larger error that would need to be
addressed.